Split server and client implementations
